feat: create BaseComponentCommits component and display it when user hovers over commits
