+++ ui/blocks/src/ComponentCommits/ComponentCommits.tsx
@@ -1,24 +1,21 @@
 /* eslint-disable react/display-name */
 /** @jsx jsx */
-import { FC, useMemo } from 'react';
+import { FC } from 'react';
 import { jsx } from 'theme-ui';
-import { Column } from 'react-table';
 import {
   BlockContainer,
   BlockContainerProps,
-  Table,
 } from '@component-controls/components';
+import { StoryInputProps, useStoryComponent } from '@component-controls/store';
 import {
-  StoryInputProps,
-  useConfig,
-  useStoryComponent,
-} from '@component-controls/store';
-import { Commit, Component, dateToLocalString } from '@component-controls/core';
-import { GithubAvatar } from '@component-controls/components';
+  BaseComponentCommits,
+  BaseComponentCommitsProps,
+} from './BaseComponentCommits';
 import { useCustomProps } from '../context';
 
 export type ComponentCommitsProps = BlockContainerProps &
-  StoryInputProps & { storeComponent?: Component };
+  StoryInputProps &
+  Pick<BaseComponentCommitsProps, 'pagination'>;
 
 /**
  * Displays commit history for a component
@@ -26,97 +23,21 @@ export type ComponentCommitsProps = BlockContainerProps &
 export const ComponentCommits: FC<ComponentCommitsProps> = ({
   id,
   name,
-  storeComponent,
+  pagination,
   ...rest
 }) => {
   const props = useCustomProps<ComponentCommitsProps>('commits', rest);
-  const component = storeComponent
-    ? storeComponent
-    : useStoryComponent({ id, name });
-
-  const config = useConfig();
-  const { tokens } = config;
+  const component = useStoryComponent({ id, name });
 
   console.log(component);
-
-  const columns = useMemo(
-    () =>
-      [
-        {
-          Header: 'Date',
-          accessor: 'authorDate',
-          Cell: ({
-            row: {
-              original: { authorDate },
-            },
-          }) => (
-            <span>
-              {authorDate ? dateToLocalString(new Date(authorDate)) : 'N/A'}
-            </span>
-          ),
-        },
-        {
-          Header: 'Author',
-          accessor: 'authorName',
-          Cell: ({
-            row: {
-              original: { authorName },
-            },
-          }) => {
-            return authorName ? (
-              <div style={{ display: 'flex' }}>
-                <p style={{ paddingRight: '10px' }}>{authorName}</p>
-                <GithubAvatar
-                  username={authorName}
-                  size={22}
-                  githubAccessToken={tokens?.githubAccessToken}
-                />
-              </div>
-            ) : null;
-          },
-        },
-        {
-          Header: 'Commit Message',
-          accessor: 'subject',
-          Cell: ({
-            row: {
-              original: { subject },
-            },
-          }) => <p>{subject}</p>,
-        },
-      ] as Column<Commit>[],
-    [],
-  );
 
   if (!component?.fileInfo?.commits) {
     return null;
   }
 
-  const paginationProps = {
-    pageIndex: 0,
-    pageSize: 10,
-    pageTemplate: 'Page ${pageIndex} of ${pageLength}',
-    pageVisible: true,
-    pageSizeTemplate: '${pageSize} rows',
-    pageSizeVisible: true,
-    goToPageVisible: true,
-    goToPageTemplate: 'Go to page:',
-  };
-
   return (
     <BlockContainer {...props}>
-      <Table<Commit>
-        sorting={true}
-        data={component.fileInfo.commits}
-        columns={columns}
-        sortBy={[
-          {
-            id: 'authorDate',
-            desc: true,
-          },
-        ]}
-        pagination={paginationProps}
-      />
+      <BaseComponentCommits component={component} pagination={pagination} />
     </BlockContainer>
   );
 };
