+++ ui/blocks/src/ComponentCommits/index.ts
@@ -1,1 +1,2 @@
+export * from './BaseComponentCommits';
 export * from './ComponentCommits';
