+++ ui/blocks/src/test/storyStore.tsx
@@ -340,6 +340,60 @@ and a [link](https://google.com)
             authorDate: '2021-02-21T22:48:26.000Z',
             subject: 'first commit!',
           },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
+          {
+            committerName: 'atanasster',
+            authorName: 'atanasster',
+            authorDate: '2021-02-21T22:48:26.000Z',
+            subject: 'first commit!',
+          },
         ],
       },
       info: {
